backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==4.6.0
mongoengine==0.28.2
dnspython==2.4.2
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
pytz==2023.3
requests==2.31.0
redis==5.0.1

